Watchlist Modal styling and functionality
Styled the modal and added the ability to clear the list

+++ src/components/RemoveMovieButton.jsx
@@ -2,14 +2,18 @@ import React from "react";
 
 export default function RemoveMovieButton({ title, update }) {
   function handleClick() {
-    let parsedList = JSON.parse(localStorage.getItem("MovieHoundWatchList"));
-    let filteredList = parsedList.filter((movie) => {
-      if (movie[0] != title) {
-        return movie;
-      }
-    });
-    localStorage.setItem("MovieHoundWatchList", JSON.stringify(filteredList));
-    update(filteredList);
+    let confirmation = confirm(`Remove '${title}' from your Watch List?`);
+
+    if (confirmation) {
+      let parsedList = JSON.parse(localStorage.getItem("MovieHoundWatchList"));
+      let filteredList = parsedList.filter((movie) => {
+        if (movie[0] != title) {
+          return movie;
+        }
+      });
+      localStorage.setItem("MovieHoundWatchList", JSON.stringify(filteredList));
+      update(filteredList);
+    }
   }
 
   return <button onClick={handleClick}>×</button>;
